refactoring and moving builtin to builtin.c
