+++ parse.c
@@ -294,8 +294,8 @@ static Node *new_binary(NodeKind kind, Node *lhs, Node *rhs, Token *tok)
   Node *node = new_node(kind, tok);
   if (isDotfile && dotf != NULL)
   {
-    fprintf(dotf, "%s%d -> %s%d\n", nodekind2str(node->kind), node->unique_number, nodekind2str(rhs->kind), rhs->unique_number);
-    fprintf(dotf, "%s%d -> %s%d\n", nodekind2str(node->kind), node->unique_number, nodekind2str(lhs->kind), lhs->unique_number);
+    fprintf(dotf, "%s%d -> %s%d\n", kNodeKindStr[node->kind], node->unique_number, kNodeKindStr[rhs->kind], rhs->unique_number);
+    fprintf(dotf, "%s%d -> %s%d\n", kNodeKindStr[node->kind], node->unique_number, kNodeKindStr[lhs->kind], lhs->unique_number);
   }
 
   node->lhs = lhs;
@@ -315,7 +315,7 @@ static Node *new_unary(NodeKind kind, Node *expr, Token *tok)
   Node *node = new_node(kind, tok);
 
   if (isDotfile && dotf != NULL)
-    fprintf(dotf, "%s%d -> %s%d\n", nodekind2str(node->kind), node->unique_number, nodekind2str(expr->kind), expr->unique_number);
+    fprintf(dotf, "%s%d -> %s%d\n", kNodeKindStr[node->kind], node->unique_number, kNodeKindStr[expr->kind], expr->unique_number);
   node->lhs = expr;
   add_type(node->lhs);
   return node;
@@ -2720,19 +2720,19 @@ static Node *stmt(Token **rest, Token *tok, bool chained)
     node->cond = to_bool(expr(&tok, tok));
 
     if (isDotfile && dotf != NULL)
-      fprintf(dotf, "%s%d -> %s%d\n", nodekind2str(node->kind), node->unique_number, nodekind2str(node->cond->kind), node->cond->unique_number);
+      fprintf(dotf, "%s%d -> %s%d\n", kNodeKindStr[node->kind], node->unique_number, kNodeKindStr[node->cond->kind], node->cond->unique_number);
 
     SET_CTX(ctx);        
     tok = skip(tok, ")", ctx);
     node->then = stmt(&tok, tok, true);
 
     if (isDotfile && dotf != NULL)
-      fprintf(dotf, "%s%d -> %s%d\n", nodekind2str(node->kind), node->unique_number, nodekind2str(node->then->kind), node->then->unique_number);
+      fprintf(dotf, "%s%d -> %s%d\n", kNodeKindStr[node->kind], node->unique_number, kNodeKindStr[node->then->kind], node->then->unique_number);
     if (equal(tok, "else"))
     {
       node->els = stmt(&tok, tok->next, true);
       if (isDotfile && dotf != NULL)
-        fprintf(dotf, "%s%d -> %s%d\n", nodekind2str(node->kind), node->unique_number, nodekind2str(node->els->kind), node->els->unique_number);
+        fprintf(dotf, "%s%d -> %s%d\n", kNodeKindStr[node->kind], node->unique_number, kNodeKindStr[node->els->kind], node->els->unique_number);
     }
 
     if (is_const_expr(node->cond)) {
@@ -2900,7 +2900,7 @@ static Node *stmt(Token **rest, Token *tok, bool chained)
       node->cond = expr(&tok, tok);
 
       if (isDotfile && dotf != NULL)
-        fprintf(dotf, "%s%d -> %s%d\n", nodekind2str(node->kind), node->unique_number, nodekind2str(node->cond->kind), node->cond->unique_number);
+        fprintf(dotf, "%s%d -> %s%d\n", kNodeKindStr[node->kind], node->unique_number, kNodeKindStr[node->cond->kind], node->cond->unique_number);
     }
     SET_CTX(ctx);      
     tok = skip(tok, ";", ctx);
@@ -2910,7 +2910,7 @@ static Node *stmt(Token **rest, Token *tok, bool chained)
       node->inc = expr(&tok, tok);
 
       if (isDotfile && dotf != NULL)
-        fprintf(dotf, "%s%d -> %s%d\n", nodekind2str(node->kind), node->unique_number, nodekind2str(node->inc->kind), node->inc->unique_number);
+        fprintf(dotf, "%s%d -> %s%d\n", kNodeKindStr[node->kind], node->unique_number, kNodeKindStr[node->inc->kind], node->inc->unique_number);
     }
     SET_CTX(ctx);     
     tok = skip(tok, ")", ctx);
@@ -2918,7 +2918,7 @@ static Node *stmt(Token **rest, Token *tok, bool chained)
     node->then = stmt(rest, tok, true);
 
     if (isDotfile && dotf != NULL)
-      fprintf(dotf, "%s%d -> %s%d\n", nodekind2str(node->kind), node->unique_number, nodekind2str(node->then->kind), node->then->unique_number);
+      fprintf(dotf, "%s%d -> %s%d\n", kNodeKindStr[node->kind], node->unique_number, kNodeKindStr[node->then->kind], node->then->unique_number);
     leave_scope();
     brk_label = brk;
     cont_label = cont;
@@ -3099,7 +3099,7 @@ static Node *compound_stmt(Token **rest, Token *tok, Node **last)
   if (isDotfile && dotf != NULL)
   {
     if (node->body != NULL)
-      fprintf(dotf, "%s%d -> %s%d\n", nodekind2str(node->kind), node->unique_number, nodekind2str(node->body->kind), node->body->unique_number);
+      fprintf(dotf, "%s%d -> %s%d\n", kNodeKindStr[node->kind], node->unique_number, kNodeKindStr[node->body->kind], node->body->unique_number);
   }
   *rest = tok->next;
   return node;
@@ -3163,7 +3163,7 @@ static Node *compound_stmt2(Token **rest, Token *tok)
   if (isDotfile && dotf != NULL)
   {
     if (node->body != NULL)
-      fprintf(dotf, "%s%d -> %s%d\n", nodekind2str(node->kind), node->unique_number, nodekind2str(node->body->kind), node->body->unique_number);
+      fprintf(dotf, "%s%d -> %s%d\n", kNodeKindStr[node->kind], node->unique_number, kNodeKindStr[node->body->kind], node->body->unique_number);
   }
   *rest = tok;
   return node;
@@ -3183,7 +3183,7 @@ static Node *expr_stmt(Token **rest, Token *tok)
   node->lhs = expr(&tok, tok);
 
   if (isDotfile && dotf != NULL)
-    fprintf(dotf, "%s%d -> %s%d\n", nodekind2str(node->kind), node->unique_number, nodekind2str(node->lhs->kind), node->lhs->unique_number);
+    fprintf(dotf, "%s%d -> %s%d\n", kNodeKindStr[node->kind], node->unique_number, kNodeKindStr[node->lhs->kind], node->lhs->unique_number);
 
   SET_CTX(ctx);     
   *rest = skip(tok, ";", ctx);
@@ -8388,563 +8388,6 @@ Obj *parse(Token *tok)
   scan_globals();
 
   return globals;
-}
-
-// for dot diagram
-char *nodekind2str(NodeKind kind)
-{
-  switch (kind)
-  {
-  case ND_NULL_EXPR: return "NULL_EXPR";
-  case ND_ADD: return "ADD";
-  case ND_SUB: return "SUB";
-  case ND_MUL: return "MUL";
-  case ND_DIV: return "DIV";
-  case ND_POS: return "POS";
-  case ND_NEG: return "NEG";
-  case ND_MOD: return "MOD";
-  case ND_BITAND: return "BITAND";
-  case ND_BITOR: return "BITOR";
-  case ND_BITXOR: return "BITXOR";
-  case ND_SHL: return "SHL";
-  case ND_SHR: return "SHR";
-  case ND_EQ: return "EQ";
-  case ND_NE: return "NE";
-  case ND_LT: return "LT";
-  case ND_LE: return "LE";
-  case ND_ASSIGN: return "ASSIGN";
-  case ND_COND: return "COND"; 
-  case ND_COMMA: return "COMMA";
-  case ND_MEMBER: return "MEMBER"; 
-  case ND_ADDR: return "ADDR";
-  case ND_DEREF: return "DEREF";
-  case ND_NOT: return "NOT";
-  case ND_BITNOT: return "BITNOT"; 
-  case ND_LOGAND: return "LOGAND"; 
-  case ND_LOGOR: return "LOGOR"; 
-  case ND_RETURN: return "RETURN";
-  case ND_IF: return "IF";
-  case ND_FOR: return "FOR"; 
-  case ND_DO: return "DO"; 
-  case ND_SWITCH: return "SWITCH"; 
-  case ND_CASE: return "CASE"; 
-  case ND_BLOCK: return "BLOCK"; 
-  case ND_GOTO: return "GOTO"; 
-  case ND_GOTO_EXPR: return "GOTO_EXPR"; 
-  case ND_LABEL: return "LABEL"; 
-  case ND_LABEL_VAL: return "LABEL_VAL";
-  case ND_FUNCALL: return "FUNCALL";
-  case ND_EXPR_STMT: return "EXPRSTMR";
-  case ND_STMT_EXPR: return "STMTEXPR"; 
-  case ND_VAR: return "VAR"; 
-  case ND_VLA_PTR: return "VLAPTR"; 
-  case ND_NUM: return "NUM";
-  case ND_CAST: return "CAST";
-  case ND_MEMZERO: return "MEMZERO"; 
-  case ND_ASM: return "ASM";
-  case ND_CAS: return "CAS";
-  case ND_CAS_N: return "CAS_N";
-  case ND_EXCH: return "EXCHANGE";
-  case ND_EXCH_N: return "EXCHANGE_N";
-  case ND_CMPEXCH: return "COMPARE_EXCHANGE";
-  case ND_CMPEXCH_N: return "COMPARE_EXCHANGE_N";
-  case ND_LOAD: return "LOAD";
-  case ND_LOAD_N: return "LOAD_N";
-  case ND_STORE: return "STORE"; 
-  case ND_STORE_N: return "STORE_N"; 
-  case ND_TESTANDSET: return "TESTANDSET";
-  case ND_TESTANDSETA: return "TESTANDSETA";
-  case ND_CLEAR: return "CLEAR"; 
-  case ND_RELEASE: return "RELEASE"; 
-  case ND_FETCHADD: return "FETCHADD";
-  case ND_FETCHSUB: return "FETCHSUB";
-  case ND_FETCHXOR: return "FETCHXOR";
-  case ND_FETCHAND: return "FETCHAND";    
-  case ND_FETCHOR: return "FETCHOR";
-  case ND_SUBFETCH: return "SUBFETCH";
-  case ND_SYNC: return "SYNC";   
-  case ND_MEMBARRIER: return "MEMBARRIER"; 
-  case ND_BUILTIN_MEMCPY: return "MEMCPY";  
-  case ND_BUILTIN_MEMSET: return "MEMSET";  
-  case ND_BUILTIN_CLZ: return "CLZ";   
-  case ND_BUILTIN_CLZL: return "CLZL";   
-  case ND_BUILTIN_CLZLL: return "CLZLL";  
-  case ND_BUILTIN_CTZ: return "CTZ";   
-  case ND_BUILTIN_CTZL: return "CTZL";  
-  case ND_BUILTIN_CTZLL: return "CTZLL";  
-  case ND_BUILTIN_INFF: return "INFF";   
-  case ND_BUILTIN_INF: return "INFF";  
-  case ND_BUILTIN_NAN: return "NAN";   
-  case ND_BUILTIN_NANF: return "NANF";   
-  case ND_BUILTIN_NANL: return "NANL";   
-  case ND_BUILTIN_ISNAN: return "ISNAN"; 
-  case ND_BUILTIN_HUGE_VAL: return "ISNAN"; 
-  case ND_BUILTIN_HUGE_VALF: return "ISNAN"; 
-  case ND_BUILTIN_HUGE_VALL: return "ISNAN";
-  case ND_POPCOUNT: return "POPCOUNT";
-  case ND_POPCOUNTL: return "POPCOUNTL";
-  case ND_POPCOUNTLL: return "POPCOUNTLL";
-  case ND_RETURN_ADDR: return "RETURN_ADDRESS";  
-  case ND_BUILTIN_FRAME_ADDRESS: return "FRAME_ADDRESS"; 
-  case ND_STDC_BIT_CEIL: return "STDC_BIT_CEIL"; 
-  case ND_BUILTIN_ADD_OVERFLOW: return "ADD_OVERFLOW";    
-  case ND_BUILTIN_SUB_OVERFLOW: return "SUB_OVERFLOW";    
-  case ND_BUILTIN_MUL_OVERFLOW: return "MUL_OVERFLOW"; 
-  case ND_UADD_OVERFLOW: return "UADD_OVERFLOW";  
-  case ND_UADDL_OVERFLOW: return "UADDL_OVERFLOW";    
-  case ND_UADDLL_OVERFLOW: return "UADDLL_OVERFLOW";    
-  case ND_UMUL_OVERFLOW: return "UMUL_OVERFLOW";  
-  case ND_UMULL_OVERFLOW: return "UMULL_OVERFLOW";    
-  case ND_UMULLL_OVERFLOW: return "UMULLL_OVERFLOW";      
-  case ND_BUILTIN_BSWAP16:  return "BSWAP16";   
-  case ND_BUILTIN_BSWAP32: return "BSWAP32";    
-  case ND_BUILTIN_BSWAP64: return "BSWAP64"; 
-  case ND_ALLOC: return "ALLOCA";
-  case ND_ABORT: return "ABORT";
-  case ND_EXPECT: return "EXPECT";
-  case ND_EMMS: return "EMMS";
-  case ND_LFENCE: return "LFENCE";
-  case ND_MFENCE: return "MFENCE";
-  case ND_SFENCE: return "SFENCE";
-  case ND_PAUSE: return "PAUSE";
-  case ND_STMXCSR: return "STMXCSR";
-  case ND_LDMXCSR: return "LDMXCSR";    
-  case ND_CVTPI2PS: return "CVTPI2PS";    
-  case ND_CVTPS2PI: return "CVTPS2PI";
-  case ND_CLFLUSH: return "CLFLUSH"; 
-  case ND_VECINITV2SI: return "VEC_INIT_V2SI";
-  case ND_VECEXTV2SI: return "VEC_EXT_V2SI";
-  case ND_VECEXTV4SI: return "VEC_EXT_V4SI";  
-  case ND_VECINITV8QI: return "VEC_INIT_V8QI";   
-  case ND_VECINITV4HI: return "VEC_INIT_V4HI";        
-  case ND_PACKSSWB: return "PACKSSWB";
-  case ND_PACKSSDW: return "PACKSSDW";
-  case ND_PACKUSWB: return "PACKUSWB";
-  case ND_PUNPCKHBW: return "PUNPCKHBW";
-  case ND_PUNPCKHWD: return "PUNPCKHWD";   
-  case ND_PUNPCKHDQ: return "PUNPCKHDQ";   
-  case ND_PUNPCKLBW: return "PUNPCKLBW";  
-  case ND_PUNPCKLWD: return "PUNPCKLWD";   
-  case ND_PUNPCKLDQ: return "PUNPCKLDQ";   
-  case ND_PADDB: return "PADDB";   
-  case ND_PADDW: return "PADDW";  
-  case ND_PADDD: return "PADDD";
-  case ND_PADDQ: return "PADDQ";   
-  case ND_PADDSB: return "PADDSB";  
-  case ND_PADDSW: return "PADDSW";   
-  case ND_PADDUSB: return "PADDUSB"; 
-  case ND_PADDUSW: return "PADDUSW";   
-  case ND_PSUBB: return "PSUBB";    
-  case ND_PSUBW: return "PSUBW";   
-  case ND_PSUBD: return "PSUBD";     
-  case ND_PSUBQ: return "PSUBQ";  
-  case ND_PSUBSB: return "PSUBSB";   
-  case ND_PSUBSW: return "PSUBSW"; 
-  case ND_PSUBUSB: return "PSUBUSB"; 
-  case ND_PSUBUSW: return "PSUBUSW";     
-  case ND_PMADDWD: return "PMADDWD";    
-  case ND_PMULHW: return "PMULHW";   
-  case ND_PMULLW: return "PMULLW";  
-  case ND_PSLLW: return "PSLLW";   
-  case ND_PSLLWI: return "PSLLWI"; 
-  case ND_PSLLD: return "PSLLD";      
-  case ND_PSLLDI: return "PSLLDI";   
-  case ND_PSLLQ: return "PSLLQ";
-  case ND_PSLLQI: return "PSLLQI";  
-  case ND_PSRAW: return "PSRAW";   
-  case ND_PSRAWI: return "PSRAWI";
-  case ND_PSRAD: return "PSRAD";  
-  case ND_PSRADI: return "PSRADI";
-  case ND_PSRLW: return "PSRLW";      
-  case ND_PSRLWI: return "PSRLWI";   
-  case ND_PSRLD: return "PSRLD";      
-  case ND_PSRLDI: return "PSRLDI";  
-  case ND_PSRLQ: return "PSRLQ";      
-  case ND_PSRLQI: return "PSRLQI"; 
-  case ND_PAND: return "PAND";  
-  case ND_PANDN: return "PANDN";    
-  case ND_POR: return "POR";
-  case ND_PXOR: return "PXOR";    
-  case ND_PCMPEQB: return "PCMPEQB";     
-  case ND_PCMPGTB: return "PCMPGTB";  
-  case ND_PCMPEQW: return "PCMPEQW";  
-  case ND_PCMPEQD: return "PCMPEQD";        
-  case ND_PCMPGTW: return "PCMPGTW";  
-  case ND_PCMPGTD: return "PCMPGTD";  
-  case ND_ADDSS: return "ADDSS";  
-  case ND_SUBSS: return "SUBSS";   
-  case ND_MULSS: return "MULSS";  
-  case ND_DIVSS: return "DIVSS";  
-  case ND_SQRTSS: return "SQRTSS"; 
-  case ND_SQRTPS: return "SQRTPS";     
-  case ND_RSQRTSS: return "RSQRTSS";  
-  case ND_RSQRTPS: return "RSQRTPS";         
-  case ND_RCPSS: return "RCPSS"; 
-  case ND_RCPPS: return "RCPPS";      
-  case ND_MINSS: return "MINSS";
-  case ND_MINPS: return "MINPS"; 
-  case ND_MAXPS: return "MAXPS";        
-  case ND_MAXSS: return "MAXSS"; 
-  case ND_ANDPS: return "ANDPS";       
-  case ND_ANDNPS: return "ANDNPS";   
-  case ND_ORPS: return "ORPS";  
-  case ND_XORPS: return "XORPS";   
-  case ND_CMPEQSS: return "CMPEQSS";    
-  case ND_CMPLTSS: return "CMPLTSS";  
-  case ND_CMPLESS: return "CMPLESS";    
-  case ND_MOVSS: return "MOVSS";   
-  case ND_CMPNEQSS: return "CMPNEQSS"; 
-  case ND_CMPNLTSS: return "CMPNLTSS"; 
-  case ND_CMPNLESS: return "CMPNLESS"; 
-  case ND_CMPORDSS: return "CMPORDSS"; 
-  case ND_CMPUNORDSS: return "CMPUNORDSS";    
-  case ND_CMPEQPS: return "CMPEQPS";  
-  case ND_CMPLTPS: return "CMPLTPS";     
-  case ND_CMPLEPS: return "CMPLEPS";  
-  case ND_CMPGTPS: return "CMPGTPS";     
-  case ND_CMPGEPS: return "CMPGEPS"; 
-  case ND_CMPNEQPS: return "CMPNEQPS";  
-  case ND_CMPNLTPS: return "CMPNLTPS";         
-  case ND_CMPNLEPS: return "CMPNLEPS"; 
-  case ND_CMPNGTPS: return "CMPNGTPS";     
-  case ND_CMPNGEPS: return "CMPNGEPS"; 
-  case ND_CMPORDPS: return "CMPORDPS";   
-  case ND_CMPUNORDPS: return "CMPUNORDPS";  
-  case ND_COMIEQ: return "COMIEQ";  
-  case ND_COMILT: return "COMILT";     
-  case ND_COMILE: return "COMILE"; 
-  case ND_COMIGT: return "COMIGT"; 
-  case ND_COMIGE: return "COMIGE";     
-  case ND_COMINEQ: return "COMINEQ";                                                                                                                                                                                                                                                                                                                             
-  case ND_UCOMIEQ: return "UCOMIEQ";  
-  case ND_UCOMILT: return "UCOMILT";     
-  case ND_UCOMILE: return "UCOMILE"; 
-  case ND_UCOMIGT: return "UCOMIGT"; 
-  case ND_UCOMIGE: return "UCOMIGE";     
-  case ND_UCOMINEQ: return "UCOMINEQ";
-  case ND_CVTSS2SI: return "CVTSS2SI";    
-  case ND_CVTSS2SI64: return "CVTSS2SI64";   
-  case ND_CVTTSS2SI: return "CVTTSS2SI";  
-  case ND_CVTTSS2SI64: return "CVTTSS2SI64";
-  case ND_CVTTPS2PI: return "CVTTPS2PI";   
-  case ND_CVTSI2SS: return "CVTSI2SS";   
-  case ND_CVTSI642SS: return "CVTSI642SS";  
-  case ND_MOVLHPS: return "MOVLHPS";    
-  case ND_MOVHLPS: return "MOVHLPS"; 
-  case ND_UNPCKHPS: return "UNPCKHPS";  
-  case ND_UNPCKLPS: return "UNPCKLPS";  
-  case ND_LOADHPS: return "LOADHPS";        
-  case ND_STOREHPS: return "STOREHPS"; 
-  case ND_LOADLPS: return "LOADLPS";        
-  case ND_STORELPS: return "STORELPS"; 
-  case ND_MOVMSKPS: return "MOVMSKPS";    
-  case ND_SHUFPS: return "SHUFPS";       
-  case ND_SHUFFLE: return "SHUFFLE";     
-  case ND_PMAXSW: return "PMAXSW";     
-  case ND_PMAXUB: return "PMAXUB";   
-  case ND_PMINSW: return "PMINSW"; 
-  case ND_PMINUB: return "PMINUB";        
-  case ND_PMOVMSKB: return "PMOVMSKB";
-  case ND_PMULHUW: return "PMULHUW";   
-  case ND_MASKMOVQ: return "MASKMOVQ";   
-  case ND_PAVGB: return "PAVGB";            
-  case ND_PAVGW: return "PAVGW"; 
-  case ND_PSADBW: return "PSADBW";    
-  case ND_MOVNTQ: return "MOVNTQ";    
-  case ND_MOVNTPS: return "MOVNTPS";         
-  case ND_SHUFPD: return "SHUFPD";       
-  case ND_ADDSD: return "ADDSD";   
-  case ND_SUBSD: return "SUBSD";       
-  case ND_MULSD: return "MULSD";   
-  case ND_DIVSD: return "DIVSD";    
-  case ND_SQRTPD: return "SQRTPD"; 
-  case ND_SQRTSD: return "SQRTSD"; 
-  case ND_MOVSD: return "MOVSD";        
-  case ND_MINPD: return "MINPD";            
-  case ND_MINSD: return "MINSD";        
-  case ND_MAXPD: return "MAXPD";  
-  case ND_MAXSD: return "MAXSD";  
-  case ND_ANDPD: return "ANDPD";    
-  case ND_ANDNPD: return "ANDNPD";  
-  case ND_ORPD: return "ORPD";         
-  case ND_XORPD: return "XORPD";      
-  case ND_CMPEQPD: return "CMPEQPD";    
-  case ND_CMPLTPD: return "CMPLTPD"; 
-  case ND_CMPLEPD: return "CMPLEPD"; 
-  case ND_CMPGTPD: return "CMPGTPD"; 
-  case ND_CMPGEPD: return "CMPGEPD";  
-  case ND_CMPNEQPD: return "CMPNEQPD"; 
-  case ND_CMPNLTPD: return "CMPNLTPD";  
-  case ND_CMPNLEPD: return "CMPNLEPD";  
-  case ND_CMPNGTPD: return "CMPNGTPD";   
-  case ND_CMPNGEPD: return "CMPNGEPD";   
-  case ND_CMPORDPD: return "CMPORDPD";   
-  case ND_CMPUNORDPD: return "CMPUNORDPD";
-  case ND_CMPEQSD: return "CMPEQSD";   
-  case ND_CMPLTSD: return "CMPLTSD";    
-  case ND_CMPLESD: return "CMPLESD";    
-  case ND_CMPNEQSD: return "CMPNEQSD";   
-  case ND_CMPNLTSD: return "CMPNLTSD";     
-  case ND_CMPNLESD: return "CMPNLESD";  
-  case ND_CMPORDSD: return "CMPORDSD";    
-  case ND_CMPUNORDSD: return "CMPUNORDSD"; 
-  case ND_COMISDEQ: return "COMISDEQ"; 
-  case ND_COMISDLT: return "COMISDLT"; 
-  case ND_COMISDLE: return "COMISDLE"; 
-  case ND_COMISDGT: return "COMISDGT"; 
-  case ND_COMISDGE: return "COMISDGE"; 
-  case ND_COMISDNEQ: return "COMISDNEQ"; 
-  case ND_UCOMISDEQ: return "UCOMISDEQ"; 
-  case ND_UCOMISDLT: return "UCOMISDLT"; 
-  case ND_UCOMISDLE: return "UCOMISDLE"; 
-  case ND_UCOMISDGT: return "UCOMISDGT"; 
-  case ND_UCOMISDGE: return "UCOMISDGE"; 
-  case ND_UCOMISDNEQ: return "UCOMISDNEQ"; 
-  case ND_MOVQ128: return "MOVQ128"; 
-  case ND_CVTDQ2PD: return "CVTDQ2PD"; 
-  case ND_CVTDQ2PS: return "CVTDQ2PS"; 
-  case ND_CVTPD2DQ: return "CVTPD2DQ"; 
-  case ND_CVTPD2PI: return "CVTPD2PI"; 
-  case ND_CVTPD2PS: return "CVTPD2PS"; 
-  case ND_CVTTPD2DQ: return "CVTTPD2DQ";
-  case ND_CVTTPD2PI: return "CVTTPD2PI";  
-  case ND_CVTPI2PD: return "CVTPI2PD"; 
-  case ND_CVTPS2DQ: return "CVTPS2DQ";  
-  case ND_CVTTPS2DQ: return "CVTTPS2DQ";
-  case ND_CVTPS2PD: return "CVTPS2PD"; 
-  case ND_CVTSD2SI: return "CVTSD2SI"; 
-  case ND_CVTSD2SI64: return "CVTSD2SI64";
-  case ND_CVTTSD2SI: return "CVTTSD2SI";  
-  case ND_CVTTSD2SI64: return "CVTTSD2SI64";
-  case ND_CVTSD2SS: return "CVTSD2SS"; 
-  case ND_CVTSI2SD: return "CVTSI2SD"; 
-  case ND_CVTSI642SD: return "CVTSI642SD";
-  case ND_CVTSS2SD: return "CVTSS2SD";  
-  case ND_UNPCKHPD: return "UNPCKHPD";  
-  case ND_UNPCKLPD: return "UNPCKLPD";  
-  case ND_LOADHPD: return "LOADHPD";       
-  case ND_LOADLPD: return "LOADLPD";  
-  case ND_MOVMSKPD: return "MOVMSKPD";    
-  case ND_PACKSSWB128: return "PACKSSWB128";   
-  case ND_PACKSSDW128: return "PACKSSDW128";   
-  case ND_PACKUSWB128: return "PACKUSWB128";     
-  case ND_PUNPCKHBW128: return "PUNPCKHB128";   
-  case ND_PUNPCKHWD128: return "PUNPCKHWD128";  
-  case ND_PUNPCKHDQ128: return "PUNPCKHDQ128";   
-  case ND_PUNPCKHQDQ128: return "PUNPCKHQDQ128"; 
-  case ND_PUNPCKLBW128: return "PUNPCKLBW128";   
-  case ND_PUNPCKLWD128: return "PUNPCKLWD128";   
-  case ND_PUNPCKLDQ128: return "PUNPCKLDQ128";      
-  case ND_PUNPCKLQDQ128: return "PUNPCKLQDQ128";      
-  case ND_PADDSB128: return "PADDSB128";     
-  case ND_PADDSW128: return "PADDSW128";   
-  case ND_PADDUSB128: return "PADDUSB128";      
-  case ND_PADDUSW128: return "PADDUSW128";   
-  case ND_PSUBSB128: return "PSUBSB128";      
-  case ND_PSUBSW128: return "PSUBSW128";  
-  case ND_PSUBUSB128: return "PSUBUSB128";   
-  case ND_PSUBUSW128: return "PSUBUSW128";    
-  case ND_PMADDWD128: return "PMADDWD128";         
-  case ND_PMULHW128: return "PMULHW128";  
-  case ND_PMULUDQ: return "PMULUDQ";    
-  case ND_PMULUDQ128: return "PMULUDQ128";      
-  case ND_PSLLWI128: return "PSLLWI128";           
-  case ND_PSLLDI128: return "PSLLDI128";   
-  case ND_PSLLQI128: return "PSLLQI128"; 
-  case ND_PSRAWI128: return "PSRAWI128";     
-  case ND_PSRADI128: return "PSRADI128";   
-  case ND_PSRLWI128: return "PSRLWI128";     
-  case ND_PSRLDI128: return "PSRLDI128"; 
-  case ND_PSRLQI128: return "PSRLQI128";   
-  case ND_PSLLW128: return "PSLLW128";      
-  case ND_PSLLD128: return "PSLLD128";   
-  case ND_PSLLQ128: return "PSLLQ128";     
-  case ND_PSRAW128: return "PSRAW128";   
-  case ND_PSRAD128: return "PSRAD128";         
-  case ND_PSRLW128: return "PSRLW128";   
-  case ND_PSRLD128: return "PSRLD128";   
-  case ND_PSRLQ128: return "PSRLQ128";  
-  case ND_PANDN128: return "PANDN128";     
-  case ND_PMAXSW128: return "PMAXSW128";         
-  case ND_PMAXUB128: return "PMAXUB128";         
-  case ND_PMINSW128: return "PMINSW128"; 
-  case ND_PMINUB128: return "PMINUB128";      
-  case ND_PMOVMSKB128: return "PMOVMSKB128";   
-  case ND_PMULHUW128: return "PMULHUW128";        
-  case ND_MASKMOVDQU: return "MASKMOVDQU";   
-  case ND_PAVGB128: return "PAVGB128";        
-  case ND_PAVGW128: return "PAVGW128";        
-  case ND_PSADBW128: return "PSADBW128";       
-  case ND_MOVNTI: return "MOVNTI";  
-  case ND_MOVNTI64: return "MOVNTI64";  
-  case ND_MOVNTDQ: return "MOVNTDQ";  
-  case ND_MOVNTPD: return "MOVNTPD";     
-  case ND_PARITY: return "PARITY";    
-  case ND_PARITYL: return "PARITYL";  
-  case ND_PARITYLL: return "PARITYLL";    
-  case ND_MWAIT: return "MWAIT";
-  case ND_MONITOR: return "MONITOR";
-  case ND_ADDSUBPS: return "ADDSUBPS";
-  case ND_HADDPS: return "HADDPS";
-  case ND_HSUBPS: return "HSUBPS";
-  case ND_MOVSHDUP: return "MOVSHDUP";
-  case ND_MOVSLDUP: return "MOVSLDUP";
-  case ND_ADDSUBPD: return "ADDSUBPD";
-  case ND_HADDPD: return "HADDPD";
-  case ND_HSUBPD: return "HSUBPD";
-  case ND_PHADDW128: return "PHADDW128";
-  case ND_PHADDD128: return "PHADDD128";
-  case ND_PHADDSW128: return "PHADDSW128";
-  case ND_PHADDW: return "PHADDW";
-  case ND_PHADDD: return "PHADDD";
-  case ND_PHADDSW: return "PHADDSW";
-  case ND_PHSUBW128: return "PHSUBW128";
-  case ND_PHSUBD128: return "PHSUBD128";
-  case ND_PHSUBSW128: return "PHSUBSW128";
-  case ND_PHSUBW: return "PHSUBW";
-  case ND_PHSUBD: return "PHSUBD";
-  case ND_PHSUBSW: return "PHSUBSW";
-  case ND_PMADDUBSW128: return "PMADDUBSW128";
-  case ND_PMADDUBSW: return "PMADDUBSW";
-  case ND_PMULHRSW128: return "PMULHRSW128";
-  case ND_PMULHRSW: return "PMULHRSW";
-  case ND_PSHUFB128: return "PSHUFB128";
-  case ND_PSHUFB: return "PSHUFB";
-  case ND_PSIGNB128: return "PSIGNB128";
-  case ND_PSIGNW128: return "PSIGNW128";
-  case ND_PSIGND128: return "PSIGND128";
-  case ND_PSIGNB: return "PSIGNB";
-  case ND_PSIGNW: return "PSIGNW";
-  case ND_PSIGND: return "PSIGND";
-  case ND_PABSB128: return "PABSB128";
-  case ND_PABSW128: return "PABSW128";
-  case ND_PABSD128: return "PABSD128";
-  case ND_PABSB: return "PABSB";
-  case ND_PABSW: return "PABSW";
-  case ND_PABSD: return "PABSD";
-  case ND_PTESTZ128: return "PTESTZ128";
-  case ND_PTESTC128: return "PTESTC128";
-  case ND_PTESTNZC128: return "PTESTNZC128";
-  case ND_PBLENDVB128: return "PBLENDVB128";
-  case ND_BLENDVPS: return "BLENDVPS";
-  case ND_BLENDVPD: return "BLENDVPD";
-  case ND_PMINSB128: return "PMINSB128";
-  case ND_PMAXSB128: return "PMAXSB128";
-  case ND_PMINUW128: return "PMINUW128";
-  case ND_PMAXUW128: return "PMAXUW128";
-  case ND_PMINSD128: return "PMINSD128";
-  case ND_PMAXSD128: return "PMAXSD128";  
-  case ND_PMINUD128: return "PMINUD128";
-  case ND_PMAXUD128: return "PMAXUD128";
-  case ND_PMULDQ128: return "PMULDQ128";
-  case ND_PHMINPOSUW128: return "PHMINPOSUW128";
-  case ND_PMOVSXBD128: return "PMOVSXBD128";
-  case ND_PMOVSXWD128: return "PMOVSXWD128";
-  case ND_PMOVSXBQ128: return "PMOVSXBQ128";
-  case ND_PMOVSXDQ128: return "PMOVSXDQ128";
-  case ND_PMOVSXWQ128: return "PMOVSXWQ128";
-  case ND_PMOVSXBW128: return "PMOVSXBW128";
-  case ND_PMOVZXBD128: return "PMOVZXBD128";
-  case ND_PMOVZXWD128: return "PMOVZXWD128";
-  case ND_PMOVZXBQ128: return "PMOVZXBQ128";
-  case ND_PMOVZXDQ128: return "PMOVZXDQ128";
-  case ND_PMOVZXWQ128: return "PMOVZXWQ128";
-  case ND_PMOVZXBW128: return "PMOVZXBW128";
-  case ND_PACKUSDW128: return "PACKUSDW128";
-  case ND_MOVNTDQA: return "MOVNTDQA";
-  case ND_CRC32QI: return "CRC32QI";
-  case ND_CRC32HI: return "CRC32HI";
-  case ND_CRC32SI: return "CRC32SI";
-  case ND_CRC32DI: return "CRC32DI";
-  case ND_PSHUFD: return "PSHUFD";
-  case ND_FETCHNAND: return "FETCHNAND";
-  case ND_ADD_AND_FETCH: return "ADD_AND_FETCH";
-  case ND_SUB_AND_FETCH: return "SUB_AND_FETCH";
-  case ND_BOOL_CAS: return "BOOL_CAS";
-  case ND_PREFETCH: return "PREFETCH";
-  case ND_RDTSC: return "RDTSC";
-  case ND_READEFLAGS_U64: return "READEFLAGS_U64";
-  case ND_RDSSPQ: return "RDSSPQ";
-  case ND_SAVEPREVSSP: return "SAVEPREVSSP";
-  case ND_SETSSBSY: return "SETSSBSY";
-  case ND_SLWPCB: return "SLWPCB";
-  case ND_RDPKRU: return "RDPKRU";
-  case ND_XBEGIN: return "XBEGIN";
-  case ND_XEND: return "XEND";
-  case ND_SERIALIZE: return "SERIALIZE";
-  case ND_XSUSLDTRK: return "XSUSLDTRK";
-  case ND_XRESLDTRK: return "XRESLDTRK";
-  case ND_CLUI: return "CLUI";
-  case ND_STUI: return "STUI";
-  case ND_TESTUI: return "TESTUI";
-  case ND_WBNOINVD: return "WBNOINVD";
-  case ND_XTEST: return "XTEST";
-  case ND_WBINVD: return "WBINVD";
-  case ND_RDPID: return "RDPID";
-  case ND_RDFSBASE32: return "RDFSBASE32";
-  case ND_RDFSBASE64: return "RDFSBASE64";
-  case ND_RDGSBASE32: return "RDGSBASE32";
-  case ND_RDGSBASE64: return "RDGSBASE64";
-  case ND_VZEROALL: return "VZEROALL";
-  case ND_VZEROUPPER: return "VZEROUPPER";
-  case ND_FEMMS: return "FEMMS";
-  case ND_BSRSI: return "BSRSI";
-  case ND_RDPMC: return "RDPMC";
-  case ND_RDTSCP: return "RDTSCP";
-  case ND_ROLQI: return "ROLQI";
-  case ND_ROLHI: return "ROLHI";
-  case ND_RORQI: return "RORQI";
-  case ND_RORHI: return "RORHI";
-  case ND_BSRDI: return "BSRDI";  
-  case ND_WRITEEFLAGS_U64: return "WRITEEFLAGS_U64";
-  case ND_INCSSPQ: return "INCSSPQ";
-  case ND_RSTORSSP: return "RSTORSSP";
-  case ND_WRSSD: return "WRSSD";
-  case ND_WRSSQ: return "WRSSQ";
-  case ND_WRUSSD: return "WRUSSD";
-  case ND_WRUSSQ: return "WRUSSQ";
-  case ND_CLRSSBSY: return "CLRSSBSY";
-  case ND_SBB_U32: return "SBB_U32";
-  case ND_ADDCARRYX_U32: return "ADDCARRYX_U32";
-  case ND_SBB_U64: return "SBB_U64";
-  case ND_ADDCARRYX_U64: return "ADDCARRYX_U64";  
-  case ND_TZCNT_U16: return "TZCNT_U16";
-  case ND_BEXTR_U32: return "BEXTR_U32";
-  case ND_ADDFETCH: return "ADDFETCH";
-  case ND_FPCLASSIFY: return "FPCLASSIFY";
-  case ND_ISUNORDERED: return "ISUNORDERED";
-  case ND_SIGNBIT: return "SIGNBIT";
-  case ND_SIGNBITF: return "SIGNBITF";
-  case ND_SIGNBITL: return "SIGNBITL";
-  case ND_ATOMIC_IS_LOCK_FREE: return "ATOMIC_IS_LOCK_FREE";
-  case ND_PSUBUSB256: return "PSUBUSB256";
-  case ND_PCMPGTB256_MASK: return "PCMPGTB256_MASK";
-  case ND_PSHUFB256: return "PSHUFB256";  
-  case ND_PBLENDVB256: return "PBLENDVB256";
-  case ND_PSRLDQI256: return "PSRLDQI256";
-  case ND_PSLLDQI256: return "PSLLDQI256";
-  case ND_VINSERTF128_SI256: return "VINSERTF128_SI256";    
-  case ND_SI256_SI: return "SI256_SI";
-  case ND_SI_SI256: return "SI_SI256";
-  case ND_PALIGNR256: return "PALIGNR256";
-  case ND_VPERM2I128_SI256: return "VPERM2I128_SI256";
-  case ND_PBLENDD256: return "PBLENDD256";
-  case ND_VEXTRACTF128_SI256: return "VEXTRACTF128_SI256";
-  case ND_VECEXTV16QI: return "VEC_EXT_V16QI";
-  case ND_VECEXTV8HI: return "VEC_EXT_V8HI";
-  case ND_ANDNOTSI256: return "ANDNOTSI256";
-  case ND_VECEXTV2DI: return "VEC_EXT_V2DI";
-  case ND_PMULHUW256: return "PMULHUW256";
-  case ND_PD256_PD: return "PD256_PD";
-  case ND_PS256_PS: return "PS256_PS";
-  case ND_PSRLQI256: return "PSRLQI256";
-  case ND_PSLLQI256: return "PSLLQI256";
-  case ND_PERMDI256: return "PERMDI256";
-  case ND_PSLLDI256: return "PSLLDI256";
-  case ND_PSRLDI256: return "PSRLDI256";
-  case ND_PSRADI256: return "PSRADI256";
-  default: return "UNREACHABLE"; 
-  }
 }
 
 //from COSMOPOLITAN adding function ConsumeStringLiteral
